refactor : 일관성, 폴더 일관성잇게 수정
