+++ application_project/app/src/main/java/com/alyak/detector/ui/PillDetail/components/AdditionalInfo.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.other
+package com.alyak.detector.ui.PillDetail.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
