+++ application_project/app/src/main/java/com/alyak/detector/ui/PillDetail/components/CardBox.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.PillDetail
+package com.alyak.detector.ui.PillDetail.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -72,7 +72,7 @@ import com.alyak.detector.data.dto.pill.PillDetail.MemoDTO
 import com.alyak.detector.data.dto.pill.PillDetail.SideEffectsDTO
 import com.alyak.detector.data.dto.pill.PillDetail.SpecialCautionDTO
 import com.alyak.detector.ui.components.StatusBadge
-import com.alyak.detector.ui.other.AdditionalInfoSection
+import com.alyak.detector.ui.PillDetail.components.AdditionalInfoSection
 import com.alyak.detector.ui.theme.CardBackground
 
 @Composable
