+++ application_project/app/src/main/java/com/alyak/detector/ui/search/components/SearchBar.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.search
+package com.alyak.detector.ui.search.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
